requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==8.0.1; python_version >= '3.6'
flask==2.0.1
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
werkzeug==2.0.1; python_version >= '3.6'
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23
pg8000==1.21.2
google-cloud-secret-manager==2.7.1
google-cloud-appengine-logging==0.1.4
google-cloud-logging==2.6.0
certifi==2021.5.30
charset-normalizer==2.0.6
DateTime==4.3
idna==3.2
Jinja2==3.0.1
python-dateutil==2.8.2
pytz==2021.1
pytzdata==2020.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
zope.interface==5.4.0
